+++ app/controllers/protos/protos_controller.rb
@@ -1,11 +1,13 @@
 class Protos::ProtosController < ApplicationController
+
   def show
     @proto = Proto.includes(:proto_images, :user, :tags).find(params[:id])
     @comment = Comment.new
-    @comments = @proto.comments.includes(:user)
+    @comments = @proto.comments.eager_load(:user)
     @likes = @proto.likes
-    @like = Like.where(proto_id: @proto.id, user_id: current_user.id)
+    @like = Like.where(proto_id: @proto.id, user_id: current_user.id) if user_signed_in?
   end
+
   def new
     @proto = Proto.new
     @proto.proto_images.build
@@ -26,17 +28,13 @@ class Protos::ProtosController < ApplicationController
 
   def update
     proto = Proto.find(params[:id])
-    if proto.user_id == current_user.id
-      proto.update(update_params)
-    end
+    proto.update(update_params) if proto.user_id == current_user.id
     redirect_to :root and return
   end
 
   def destroy
     proto = Proto.find(params[:id])
-    if proto.user_id == current_user.id
-      proto.destroy
-    end
+    proto.destroy if proto.user_id == current_user.id
     redirect_to :root and return
   end
